Add comments for FMI/FMU/ModelExchange differences
