+++ fmu/modelexchange.cpp
@@ -11,22 +11,26 @@ void ModelExchange::setVariable(const FMU::Variable &variable, const double valu
   variable.validate(FMU::Variable::Type::Real);
   fmu.fmi.fmi2SetReal(component, &variable.valueReference, 1, &value);
 }
+
 void ModelExchange::setVariable(const FMU::Variable &variable, const int value)
 {
   variable.validate(FMU::Variable::Type::Integer);
   fmu.fmi.fmi2SetInteger(component, &variable.valueReference, 1, &value);
 }
+
 void ModelExchange::setVariable(const FMU::Variable &variable, const char *value)
 {
   variable.validate(FMU::Variable::Type::String);
   fmu.fmi.fmi2SetString(component, &variable.valueReference, 1, &value);
 }
+
 void ModelExchange::setVariable(const FMU::Variable &variable, const bool value)
 {
   variable.validate(FMU::Variable::Type::Boolean);
   fmi2Boolean fmi2val = static_cast<fmi2Boolean>(value);
   fmu.fmi.fmi2SetBoolean(component, &variable.valueReference, 1, &fmi2val);
 }
+
 void ModelExchange::getVariable(const FMU::Variable &variable, double &value)
 {
   variable.validate(FMU::Variable::Type::Real);
@@ -34,6 +38,7 @@ void ModelExchange::getVariable(const FMU::Variable &variable, double &value)
   fmu.fmi.fmi2GetReal(component, &variable.valueReference, 1, &fmi2val);
   value = fmi2val;
 }
+
 void ModelExchange::getVariable(const FMU::Variable &variable, int &value)
 {
   variable.validate(FMU::Variable::Type::Integer);
@@ -41,6 +46,7 @@ void ModelExchange::getVariable(const FMU::Variable &variable, int &value)
   fmu.fmi.fmi2GetInteger(component, &variable.valueReference, 1, &fmi2val);
   value = fmi2val;
 }
+
 void ModelExchange::getVariable(const FMU::Variable &variable, const char *&value)
 {
   variable.validate(FMU::Variable::Type::Boolean);
@@ -48,6 +54,7 @@ void ModelExchange::getVariable(const FMU::Variable &variable, const char *&valu
   fmu.fmi.fmi2GetString(component, &variable.valueReference, 1, &fmi2val);
   value = fmi2val;
 }
+
 void ModelExchange::getVariable(const FMU::Variable &variable, bool &value)
 {
   variable.validate(FMU::Variable::Type::Boolean);
