+++ fmu/modelexchange.hpp
@@ -10,9 +10,14 @@
 
 
 namespace spawn::fmu {
+
+/// A FMU file used for modelexchange
+/// loads the FMU file, extracts the FMI dynamic library from it
+/// and initializes the FMI2 in ModelExchange mode
 class ModelExchange
 {
 public:
+  /// Todo: Validate that the provided fmu_file is valid for ModelExchange
   ModelExchange(const boost::filesystem::path &fmu_file,
                 const std::string &name,
                 const bool visible,
@@ -32,16 +37,33 @@ public:
     setVariable(fmu.getVariableByName(name), value);
   }
 
+  /// sets a fmi2Real value, throws exception if the variable type does not match
   void setVariable(const FMU::Variable &variable, const double value);
+  /// sets a fmi2Integer value, throws exception if the variable type does not match
   void setVariable(const FMU::Variable &variable, const int value);
+  /// sets a fmi2String value, throws exception if the variable type does not match
   void setVariable(const FMU::Variable &variable, const char *value);
+  /// sets a fmi2Boolean value, throws exception if the variable type does not match
   void setVariable(const FMU::Variable &variable, const bool value);
 
+  /// gets a fmi2Boolean value, throws exception if the variable type does not match
+  /// object is passed by reference to match calling semantics of the fmi2 interface
   void getVariable(const FMU::Variable &variable, double &value);
+
+  /// gets a fmi2Integer value, throws exception if the variable type does not match
+  /// object is passed by reference to match calling semantics of the fmi2 interface
   void getVariable(const FMU::Variable &variable, int &value);
+
+  /// gets a fmi2String value, throws exception if the variable type does not match
+  /// object is passed by reference to match calling semantics of the fmi2 interface
   void getVariable(const FMU::Variable &variable, const char *&value);
+
+  /// gets a fmi2Boolean value, throws exception if the variable type does not match
+  /// object is passed by reference to match calling semantics of the fmi2 interface
   void getVariable(const FMU::Variable &variable, bool &value);
 
+  /// templated getVariable for convenience that calls through to the type specific
+  /// getVariable implementations.
   template <typename VariableType>[[nodiscard]] VariableType getVariable(std::string_view name)
   {
     const auto &variable = fmu.getVariableByName(name);
@@ -50,6 +72,8 @@ public:
     return result;
   }
 
+  /// call the visitor `Func` with the appropriate type of the given variable object
+  /// semantics are similar to std::visit for Variant
   template <typename Func> auto visitVariable(Func &&func, const FMU::Variable &variable)
   {
     switch (variable.type) {
